refactor(indexer): move upsertPool heal-stage helpers to src/pool/upsert-stages.ts
Addresses two claude[bot] findings on PR #1094:

1. pool.ts grew 626 -> 805 lines, over the 600-line advisory soft cap. Move
   the seven stage helpers (healReferenceRateFeed, healPoolFees,
   resolveFeedIdAndBreakerHalt, buildMergedPool, resolveTokenDecimalsFields,
   mergeHealedDeltas, nextCreatedAt) verbatim into a new
   src/pool/upsert-stages.ts module, matching the existing src/pool/ split
   (self-heal.ts, health.ts, snapshots.ts). Only the four helpers upsertPool
   calls are exported; the three buildMergedPool sub-helpers stay private.
   DEFAULT_ORACLE_FIELDS moves with them (their signatures reference
   `Partial<typeof DEFAULT_ORACLE_FIELDS>`; keeping it in pool.ts would
   create a pool.ts <-> upsert-stages.ts cycle that the blocking no-circular
   dependency-cruiser rule forbids) and is re-exported from pool.ts so all
   existing import sites are unchanged. pool.ts: 805 -> 464 lines.

2. The comment "selfHealWrappedExchangeId above" inside buildMergedPool was
   moved verbatim from the old inline builder, where "above" pointed at the
   preceding call in upsertPool; as a standalone function it had nothing to
   anchor to. Reworded to name the upsertPool orchestrator explicitly. Same
   fix for the DEFAULT_ORACLE_FIELDS doc's "defaultPool ... directly below"
   reference, which the move orphaned the same way.

Zero logic changes: the moved block diffs empty against HEAD except the
`export` keywords and the two comment rewrites.
